move artifact integration tests to its own file
